+++ libraries/provider_x2go_client_app.rb
@@ -19,7 +19,10 @@
 #
 
 require 'chef/provider/lwrp_base'
+require_relative 'provider_x2go_client_app_debian'
+require_relative 'provider_x2go_client_app_fedora'
 require_relative 'provider_x2go_client_app_mac_os_x'
+require_relative 'provider_x2go_client_app_rhel'
 require_relative 'provider_x2go_client_app_ubuntu'
 require_relative 'provider_x2go_client_app_windows'